Add colspan support
This adds support for having colspan generated in tables. The spec is yet to be updated with GFM.

A cell with the sole contents of `-` will be ignored, and previous cells in the row will have a colspan attribute.

Side effect would be that users can't have a cell with just `-` in it.

Fixes #10